+++ src/Runtime/mmdRuntime.ts
@@ -516,10 +516,8 @@ export class MmdRuntime implements IMmdRuntime<MmdModel> {
 
         if (!this._useManualAnimationDuration && this._currentFrameTime === 0) {
             const models = this._models;
-            for (let i = 0; i < this._models.length; ++i) {
-                const model = models[i];
-                model.resetState();
-                this._needToInitializePhysicsModels.add(model);
+            for (let i = 0; i < models.length; ++i) {
+                this._needToInitializePhysicsModels.add(models[i]);
             }
 
             this._animationFrameTimeDuration = this._computeAnimationDuration();